+++ Assets/Scripts/IEntityStats.cs
@@ -3,27 +3,27 @@ using System.Collections.Generic;
 
 public interface IEntityStats
 {
-    void ChangeCurrentHealth(float value);
+    void ChangeCurrentHealth(int value);
 
-    void ChangeCurrentStamina(float value);
+    void ChangeCurrentStamina(int value);
 
-    void ChangeCurrentMana(float value);
+    void ChangeCurrentMana(int value);
 
-    void ChangePoise(float value);
+    void ChangePoise(int value);
 
     int strength { get; set; }
     int dexterity { get; set; }
     int intelligence { get; set; }
-    float maxHealth { get; set; }
-    float maxStamina { get; set; }
-    float maxMana { get; set; }
-    float currentHealth { get; set; }
-    float currentStamina { get; set; }
-    float currentMana { get; set; }
-    float healthRecovery { get; set; }
-    float staminaRecovery { get; set; }
-    float manaRecovery { get; set; }
+    int maxHealth { get; set; }
+    int maxStamina { get; set; }
+    int maxMana { get; set; }
+    int currentHealth { get; set; }
+    int currentStamina { get; set; }
+    int currentMana { get; set; }
+    int healthRecovery { get; set; }
+    int staminaRecovery { get; set; }
+    int manaRecovery { get; set; }
     int loadStage { get; set; }
-    float maxOxygen { get; set; }
-    float curOxygen { get; set; }
+    int maxOxygen { get; set; }
+    int curOxygen { get; set; }
 }
