+++ Assets/Scripts/RegenerationEffect.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class RegenerationEffect : Effect
 {
-    public override void Affect(float power, GameObject target)
+    public override void Affect(int power, GameObject target)
     {
         target.GetComponent<Creature>().ChangeCurrentHealth(power);
     }
